BACK-END-ERP/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic[email]==2.6.3
python-multipart==0.0.9
python-dotenv==1.0.0
email-validator==2.1.0
mangum==0.19.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
supabase==2.8.0
httpx==0.27.2
requests==2.32.5